Storage: add MBR partition definitions and utility functions
We are going to use this code from outside the mkfs tool, so it is
being moved from mkfs.c to a generic header file.
